// src/lib/region/vista-enhancers/types.ts
export interface VistaEnhancer {
  chromosome: string;
  start_position: number;
  end_position: number;
  organism: string;
  element_id: string;
  element_description: string;
  promoter: string;
  transgenesis: string;
  stage: string;
  expression: string;
  tissues: string;
  genome_assembly: string;
}